backend/requirements.txt
Django==2.2.12
djangorestframework==3.11.0
psycopg2==2.8.6